requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
asttokens==3.0.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.8.4
debugpy==1.8.20
decorator==5.2.1
dill==0.4.1
executing==2.2.1
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
modal==1.4.1
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pandas==3.0.1
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.11.0+cu128
torchvision==0.26.0+cu128
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
xxhash==3.6.0
yarl==1.23.0
